src/main/java/com/portfolio/website/service/PostsService.java
package com.portfolio.website.service;

import com.portfolio.website.dto.PostsSaveRequestDto;
import com.portfolio.website.repository.PostsRepository;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@AllArgsConstructor
@Service
public class PostsService {
    private PostsRepository postsRepository;

    @Transactional
    public long save(PostsSaveRequestDto dto) {
        return postsRepository.save(dto.toEntity()).getId();
    }
}
